:hammer: split helper methods in to new file and :lipstick: update waterfall graph
