+++ src/shared/nt-results/ResultsPage.react.js
@@ -3,5 +3,5 @@ import React from 'react'
 import ResultsByLoad from './ResultsByLoad.react'
 
 export default () => (
-  <ResultsByLoad sessionTypeId='58f478b8edb5f19b0ffd4655' />
+  <ResultsByLoad sessionTypeId='58f72a69edb5f19b0f04fc46' />
 )
